// circuits/utils/enums/circuitFunctions.enum.ts
export enum CircuitFunctions {
  EMPTY = "EMPTY",
  STACK_YELLOW = "STACK_YELLOW",
  STACK_RED = "STACK_RED",
  STACK_BLUE = "STACK_BLUE",
  TRANSFORM_YELLOW_RED = "TRANSFORM_YELLOW_RED",
  TRANSFORM_YELLOW_BLUE = "TRANSFORM_YELLOW_BLUE",
  TRANSFORM_RED_YELLOW = "TRANSFORM_RED_YELLOW",
  TRANSFORM_RED_BLUE = "TRANSFORM_RED_BLUE",
  TRANSFORM_BLUE_YELLOW = "TRANSFORM_BLUE_YELLOW",
  TRANSFORM_BLUE_RED = "TRANSFORM_BLUE_RED",
  TRANSFORMTWO_YELLOW_RED = "TRANSFORMTWO_YELLOW_RED",
  TRANSFORMTWO_YELLOW_BLUE = "TRANSFORMTWO_YELLOW_BLUE",
  TRANSFORMTWO_RED_YELLOW = "TRANSFORMTWO_RED_YELLOW",
  TRANSFORMTWO_RED_BLUE = "TRANSFORMTWO_RED_BLUE",
  TRANSFORMTWO_BLUE_YELLOW = "TRANSFORMTWO_BLUE_YELLOW",
  TRANSFORMTWO_BLUE_RED = "TRANSFORMTWO_BLUE_RED",
}